requirements.txt
aiogram
python-dotenv
ydb
boto3
sqlite3
pymongo